Initial setup for nested templates.
