+++ template/javascript/.editorconfig
