+++ template/javascript/.gitattributes
