+++ template/javascript/.gitignore
