+++ template/javascript/_package.json
