+++ template/javascript/cli.js
